refactor: move assess_agents to config package and clean up scoring
- Move assess_agents() from CLI module to config/__init__.py so
  diagnostics (E6) and tests can import without CLI dependency
- Use any() instead of list comprehension for action verb check
- Use search() instead of findall() for section heading detection
- Remove section separator comments
